requirements.txt
asgiref==3.5.0
Django==4.0.2
django-crispy-forms==1.14.0
Pillow==9.0.1
python-decouple==3.6
sqlparse==0.4.2
tzdata==2021.5
fontawesomefree==6.0.0
